\name{rrds-package}
\alias{rrds-package}
\alias{rrds}
\docType{package}
\title{
\packageTitle{rrds}
}
\description{
\packageDescription{rrds}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rrds}
\packageIndices{rrds}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rrds}

Maintainer: \packageMaintainer{rrds}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
